requirements.txt
langchain>=0.3.0
chromadb>=0.4.24
langchain-community>=0.3.0
langchain-experimental
langchain-google-genai
langchain-huggingface
langchain-text-splitters
pypdf
sentence-transformers
python-dotenv
streamlit
fastapi
uvicorn
python-multipart
requests